refactor: extract local LLM and video submit units
