+++ client/src/lib/index.js
@@ -55,6 +55,7 @@ export * from './slashdoCatalog.js';
 export * from './videoFinish.js';
 export * from './videoGenParams.js';
 export * from './videoGenResolutions.js';
+export * from './videoGenSubmission.js';
 export * from './videoReferenceModes.js';
 export * from './videoStagePreview.js';
 export * from './videoTilingOptions.js';
